LPS-71641 Move the classes to the proper package
